+++ preserve-ui/src/common/settings.ts
@@ -1,4 +1,5 @@
 import { LibraryGroupOption } from '../library';
+import { RepeatMode, ShuffleMode } from '../player';
 import { PlayColumn } from '../queues/columns';
 import { STORAGE_KEY_PREFIX } from './constants';
 
@@ -7,6 +8,9 @@ export const STORAGE_KEY_SETTINGS = `${STORAGE_KEY_PREFIX}_settings`;
 export class Settings {
     libraryGrouping: LibraryGroupOption;
     playlistColumns: Array<PlayColumn>;
+    repeatMode: RepeatMode;
+    shuffleMode: ShuffleMode;
+    volume: number;
 
     constructor(savedSettings: Partial<Settings>) {
         this.libraryGrouping =
@@ -18,6 +22,9 @@ export class Settings {
             PlayColumn.Artist,
             PlayColumn.Album,
         ];
+        this.shuffleMode = savedSettings.shuffleMode || ShuffleMode.Off;
+        this.repeatMode = savedSettings.repeatMode || RepeatMode.Off;
+        this.volume = savedSettings.volume || 1;
     }
 
     set<K extends keyof Settings>(k: K, v: this[K]): void {