+++ preserve-ui/tests/functional/pages/now-playing.ts
@@ -14,4 +14,22 @@ export class NowPlaying {
     static currentTime(): Cypress.Chainable<JQuery<HTMLElement>> {
         return cy.findByTestId('nowplaying-current-time');
     }
+    static pauseButton(): Cypress.Chainable<JQuery<HTMLElement>> {
+        return cy.findByLabelText('Pause');
+    }
+    static playButton(): Cypress.Chainable<JQuery<HTMLElement>> {
+        return cy.findByLabelText('Play');
+    }
+    static prevButton(): Cypress.Chainable<JQuery<HTMLElement>> {
+        return cy.findByLabelText('Previous Track');
+    }
+    static nextButton(): Cypress.Chainable<JQuery<HTMLElement>> {
+        return cy.findByLabelText('Next Track');
+    }
+    static repeatButton(): Cypress.Chainable<JQuery<HTMLElement>> {
+        return cy.findByTestId('repeat-button');
+    }
+    static shuffleButton(): Cypress.Chainable<JQuery<HTMLElement>> {
+        return cy.findByLabelText('Shuffle');
+    }
 }